Ready to test after DataMap integration

+++ chief-simulator/src/main/java/benjaminc/chief_simulator/control/Action.java
@@ -6,11 +6,11 @@ public class Action {
 
 	private ActionType type;
 	private int key;
-	private Long lastUse;
+	private Double lastUse;
 	private boolean used;
 	private boolean doOnce;
 	private boolean pressed;
-	public Action(ActionType type, int key, Long lastUse, boolean pressed, boolean used, boolean doOnce) {
+	public Action(ActionType type, int key, Double lastUse, boolean pressed, boolean used, boolean doOnce) {
 		super();
 		this.type = type;
 		this.key = key;
@@ -31,10 +31,10 @@ public class Action {
 	public void setKey(int key) { 
 		this.key = key;
 	}
-	public Long getLastUse() {
+	public Double getLastUse() {
 		return lastUse;
 	}
-	public void setLastUse(Long lastUse) {
+	public void setLastUse(Double lastUse) {
 		this.lastUse = lastUse;
 	}
 	public boolean isPressed() {
